requirements.txt
appnope==0.1.0
asgiref==3.2.3
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==3.0.3
django-tables2==2.2.1
entrypoints==0.3
enum34==1.1.6
fbpca==1.0
gcloud==0.17.0
googleapis-common-protos==1.51.0
HeapDict==1.0.1
httplib2==0.17.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-http-over-ws==0.0.6
jws==0.1.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.0
oauth2client==3.0.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.2
Pyrebase==3.0.27
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.5
python-jwt==2.0.1
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
yapf==0.29.0
